add synology changes

+++ arch/hexagon/kernel/dma.c
@@ -46,7 +46,6 @@ EXPORT_SYMBOL(dma_supported);
 
 static struct gen_pool *coherent_pool;
 
-
 /* Allocates from a pool of uncached memory that was reserved at boot time */
 
 static void *hexagon_dma_alloc_coherent(struct device *dev, size_t size,
@@ -68,7 +67,7 @@ static void *hexagon_dma_alloc_coherent(struct device *dev, size_t size,
 			panic("Can't create %s() memory pool!", __func__);
 		else
 			gen_pool_add(coherent_pool,
-				(unsigned long)pfn_to_virt(max_low_pfn),
+				pfn_to_virt(max_low_pfn),
 				hexagon_coherent_pool_size, -1);
 	}
